fix: CI — go fmt, gosec G115 suppressions, sync package-lock.json, add API Gateway custom domain
- go fmt on config, router, narrator, generator, analyst
- gosec G115: uint64->int in home.go (bounded by modulo), int->int32 in analyst.go (bounded by prompt)
- package-lock.json regenerated to include missing @emnapi deps
- API Gateway custom domain + ACM cert for api.daemoncode.app
- GitHub Actions variables set: AWS_REGION, AWS_ROLE_ARN, AWS_ACCOUNT_ID
>

+++ backend/internal/handlers/home.go
@@ -106,7 +106,7 @@ func buildHomeResponse(profile db.ShadowProfile, state *dynamo.ShadowState, patt
 	// Deterministic signal selection: hash(user_id + date) % len(quotes)
 	key := profile.UserID.String() + today
 	h := sha256.Sum256([]byte(key))
-	idx := int(binary.BigEndian.Uint64(h[:8])) % len(dailySignals)
+	idx := int(binary.BigEndian.Uint64(h[:8]) % uint64(len(dailySignals))) // #nosec G115 — safe: result bounded by len
 	if idx < 0 {
 		idx = -idx
 	}
